switched processors out and cleaned up files
