+++ src/export-processor/index.js
@@ -1,4 +1,3 @@
 module.exports = {
-    ProcessResults : require('./process-results'),
     ProcessHashes: require("./process-hashes")
 };
