+++ src/export-processor/process-hashes.js
