+++ src/processor/matching-processor.js
@@ -9,7 +9,7 @@ const logger = require('../logger/log');
 
 const dependencies = {
     Searcher: callbackify(require("../scryfall-api/index").Search.SearchList),
-    HashProcessor: require("../export-results/index").ProcessHashes,
+    HashProcessor: require("../export-processor/").ProcessHashes,
     Hash: require("../image-hashing/").Hash.HashImage
 };
